+++ ChangeLog.md
@@ -36,6 +36,7 @@
 24. Gtk2: Fix translucent background in inkscape.
 25. Gtk2: Enable translucent background in flash plugin.
 26. Improve shadow gradient.
+27. Fix chromium background color when non-custom color is used.
 
 ## 1.8.17
 1. **Add Qt5 support!!**.